refactor(core): extract shared downloadBlob into download.ts
